Corrected missing passing of Gridtype from rglmb.default function to glmbsim_NGauss.Cpp

+++ R/rglmb.R
@@ -205,7 +205,7 @@ rglmb.default<-function(n=1,y,x,mu,P,wt=1,dispersion=NULL,nu=NULL,V=NULL,family=
 #    return(start)
             
     outlist<-glmbsim_NGauss_cpp(n=n,y=y,x=x,mu=mu,P=P,offset2=offset2,wt=wt,dispersion=dispersion2,famfunc=famfunc,f1=f1,f2=f2,f3=f3,
-                                start=start,family=family$family,link=family$link)
+                                start=start,family=family$family,link=family$link,Gridtype=Gridtype)
     
   }
